+++ .gitignore
@@ -1,4 +1,5 @@
 .nyc_output
+.vscode
 coverage
 junit
 lib
