--- jest.config.jest.js
@@ -1,5 +1,0 @@
-module.exports = {
-  // eslint-disable-next-line global-require
-  ...require('./jest.config'),
-  setupFiles: ['<rootDir>/testSetupJest.js'],
-};
