+++ test/integration/commands/defineCommand.js
@@ -1,4 +1,5 @@
 import Redis from 'ioredis';
+// eslint-disable-next-line import/no-relative-parent-imports
 import { init, dispose } from '../../../src/lua';
 
 describe('defineCommand', () => {
